pyproject.toml
[project]
name = "autovideo"
version = "0.1.0"
description = "Automated daily news video generation tool with customizable LLM support"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "jinja2>=3.1.0",
    "pillow>=10.0.0",
    "edge-tts>=6.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "playwright>=1.55.0",
]
